requirements.txt
mss==10.1.0
pyqt6==6.11.0
pynput==1.8.1
pyyaml==6.0.3
opencv_python==4.13.0.92
python-uinput==1.0.1
picamera2==0.3.35